fixed minesweeper merge
