--- MinesweeperCode/Minesweeper/Minesweeper.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33403.182
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Minesweeper", "Minesweeper\Minesweeper.vcxproj", "{9DB68389-4EBB-4C6A-98DF-65B20603EF2E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9DB68389-4EBB-4C6A-98DF-65B20603EF2E}.Debug|x64.ActiveCfg = Debug|x64
-		{9DB68389-4EBB-4C6A-98DF-65B20603EF2E}.Debug|x64.Build.0 = Debug|x64
-		{9DB68389-4EBB-4C6A-98DF-65B20603EF2E}.Debug|x86.ActiveCfg = Debug|Win32
-		{9DB68389-4EBB-4C6A-98DF-65B20603EF2E}.Debug|x86.Build.0 = Debug|Win32
-		{9DB68389-4EBB-4C6A-98DF-65B20603EF2E}.Release|x64.ActiveCfg = Release|x64
-		{9DB68389-4EBB-4C6A-98DF-65B20603EF2E}.Release|x64.Build.0 = Release|x64
-		{9DB68389-4EBB-4C6A-98DF-65B20603EF2E}.Release|x86.ActiveCfg = Release|Win32
-		{9DB68389-4EBB-4C6A-98DF-65B20603EF2E}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {18826CB6-46A4-4481-B97D-F5F36148873A}
-	EndGlobalSection
-EndGlobal
