+++ MinesweeperCode/minesweeper.h
@@ -1,0 +1,7 @@
+#pragma once
+namespace MinesweeperNS {
+	class Minesweeper {
+	public:
+		Minesweeper();
+	};
+}
